+++ Modules/Intent.Modules.Application.MediatR.CRUD/CrudStrategies/GetAllImplementationStrategy.cs
@@ -32,7 +32,7 @@ namespace Intent.Modules.Application.MediatR.CRUD.CrudStrategies
 
         public bool IsMatch()
         {
-            if (_template.Model.HasDomainInteractions())
+            if (!_template.Model.HasDomainInteractions())
             {
                 return false;
             }